// cypress/model/createProductCommand.ts
/**
 * Auction.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Category } from './category';
import { SellerInformationDto } from './sellerInformationDto';


export interface CreateProductCommand { 
    productName?: string | null;
    shortDescription?: string | null;
    detailedDescription?: string | null;
    categoryType?: Category;
    startingPrice?: number;
    bidEndDate?: string;
    sellerInformation?: SellerInformationDto;
}